refactor(socket): implement singleton pattern for SocketService
